refactor: Separate utility functions into a dedicated file
