+++ src/components/CommentCard/index.jsx
@@ -1,3 +1,6 @@
+import formatDate from "../../utils/formatDate";
+import isValidBase64 from "../../utils/decodeBase64";
+
 function CommentCard({ data }) {
   const { creator, text, postDate, screenshot } = data;
   const commentCreator = creator.nickname;
@@ -22,25 +25,6 @@ function CommentCard({ data }) {
       <div className="p-2 text-gray-500 text-xs">{formattedDate}</div>
     </div>
   );
-}
-
-function formatDate(date) {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-  const hours = String(date.getHours()).padStart(2, "0");
-  const minutes = String(date.getMinutes()).padStart(2, "0");
-
-  return `${year}-${month}-${day} ${hours}:${minutes}`;
-}
-
-function isValidBase64(str) {
-  try {
-    return atob(str);
-  } catch (error) {
-    console.error(error);
-    return false;
-  }
 }
 
 export default CommentCard;
