+++ src/utils/decodeBase64.js
@@ -1,0 +1,10 @@
+function isValidBase64(str) {
+  try {
+    return atob(str);
+  } catch (error) {
+    console.error(error);
+    return false;
+  }
+}
+
+export default isValidBase64;
